+++ src/components/SearchBar/Period/Calendar/index.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useRef, useState } from "react";
+import React, { useCallback, useRef, useState, useMemo } from "react";
 
 import { DirectionType } from "_types/calendar";
 
@@ -7,23 +7,27 @@ import PREV_BUTTON from "@assets/prevButton.svg";
 import Icon from "@components/common/Icon";
 import CalendarPage from "@components/SearchBar/Period/Calendar/CalendarPage";
 import { CALENDAR_PAGE, HALF_MOVE_POINT } from "@constants/calendar";
-import { getDirectionValue } from "@utils/calendar";
+import { useCalendarState } from "@contexts/CalendarProvider";
+import { getDirectionValue, getMonthDifference } from "@utils/calendar";
 
 import * as S from "./style";
 
+// FIXME 새로고침 될 때 초기화가 되는 걸까?
 const today = new Date();
 
 const initSlideInfo: { translateX: number; direction: DirectionType } = {
   translateX: -HALF_MOVE_POINT,
   direction: null,
 };
-
+// TODO 캘린더 슬라이드 라이브러리 만들어보기
+// 슬라이더 크기, ...커스텀 스타일, 페이지,
 const Calendar = () => {
-  const todayYear = today.getFullYear();
-  const todayMonth = today.getMonth();
+  const todayYear = useMemo(() => today.getFullYear(), []);
+  const todayMonth = useMemo(() => today.getMonth(), []);
 
-  const [pageIndex, setPageIndex] = useState(0);
   const [slideInfo, setSlideInfo] = useState(initSlideInfo);
+  const { checkIn } = useCalendarState();
+  const pageIndex = useRef(getMonthDifference(today, checkIn));
   const isMovePending = useRef(false);
 
   const handleMoveCalendar = useCallback((movePoint: number, direction: DirectionType) => {
@@ -33,11 +37,11 @@ const Calendar = () => {
     }
   }, []);
 
-  const handleMoveEnd = () => {
-    setPageIndex((prev) => prev + getDirectionValue(slideInfo.direction));
+  const handleMoveEnd = useCallback(() => {
+    pageIndex.current += getDirectionValue(slideInfo.direction);
     setSlideInfo(() => initSlideInfo);
     isMovePending.current = false;
-  };
+  }, [slideInfo.direction]);
 
   const getTodayDatePassByMonth = useCallback(
     (passMonth: number) => new Date(todayYear, todayMonth + passMonth),
@@ -46,8 +50,8 @@ const Calendar = () => {
 
   const getCalendarPageArr = useCallback(
     (CALENDAR_PAGE: number) =>
-      Array.from({ length: CALENDAR_PAGE + 2 }, (_, i) => getTodayDatePassByMonth(pageIndex + i - 1)),
-    [getTodayDatePassByMonth, pageIndex],
+      Array.from({ length: CALENDAR_PAGE + 2 }, (_, i) => getTodayDatePassByMonth(pageIndex.current + i - 1)),
+    [getTodayDatePassByMonth],
   );
 
   return (
